Upgrade to WebRTC shims in Adapter.js 7.4.0
