+++ webrtc/webapp/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -39,7 +39,7 @@
     <name>webrtcAdapter</name>
     <load-group>webConferencingWebrtcGRP</load-group>
     <script>
-      <path>/js/adapter-6.2.1.js</path>
+      <path>/js/adapter-7.4.0.js</path>
     </script>
   </module>
   <module>
